requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
requests
beautifulsoup4
python-whois
tldextract
